+++ include/Covariances/CovPenta.hpp
@@ -26,7 +26,7 @@ public:
   CovPenta& operator= (const CovPenta &r);
   virtual ~CovPenta();
 
-  unsigned int getMaxNDim()   const  override { return 3; }
+  size_t getMaxNDim() const override { return 3; }
 
   String         getCovName() const override { return "Penta"; }
   int            getMinOrder() const override { return -1; }
