+++ include/Covariances/CovReg1D.hpp
@@ -30,8 +30,7 @@ public:
   int            getMinOrder() const override { return -1; }
   bool           getCompatibleSpaceR() const override { return true; }
 
-  unsigned int getMaxNDim()   const  override { return 1; }
-
+  size_t getMaxNDim() const override { return 1; }
 
 protected:
   double _evaluateCov(double h) const override;
