+++ include/Covariances/CovWendland2.hpp
@@ -30,7 +30,7 @@ public:
   CovWendland2& operator= (const CovWendland2 &r);
   virtual ~CovWendland2();
 
-  unsigned int   getMaxNDim() const  override { return 3; }
+  size_t getMaxNDim() const override { return 3; }
   String         getCovName() const override { return "Wendland-4,2"; }
   int            getMinOrder() const override { return -1; }
   bool           getCompatibleSpaceR() const override { return true; }
